Spawn logic mostly implemented for buildings. Need to do Wall and Gate logic still. Also need to use the new validation logic in the player-ghost logic.
Begin setting up post-build architecture. Currently trigger this when something is spawned, but this will be moved to the deferral pattern later on.
Refactored some remaining EmpireId uint8s into the new int32..
Greatly expanded the available functionality of IEntityInterface
Migrate FBuildingRequirements (LevelRequirements) into code instead of BPs.
and more...

+++ Source/SassC/Private/Gamemode/Sassilization/Territory/GraphNode.cpp
@@ -12,8 +12,6 @@ AGraphNode::AGraphNode()
 	PrimaryActorTick.bCanEverTick = false;
 	Connections.SetNum(8, false);
 	HasVisited.SetNum(8, false);
-	// DebugMeshComponent = CreateDefaultSubobject<UStaticMeshComponent>(TEXT("Debug Visual"));
-	// DebugMeshComponent->SetupAttachment(RootComponent);
 }
 
 // Called when the game starts or when spawned
@@ -23,7 +21,7 @@ void AGraphNode::BeginPlay()
 
 }
 
-uint8 AGraphNode::GetEmpireId() const
+int32 AGraphNode::GetEmpireId() const
 {
 	return UKismetMathLibrary::Round(this->GetAStarFScore());
 }
